Move build_single_flashblock from metering to test-utils
